Three-finger drag support.
The default for this is 1. Which probably means you won't be able to turn it off in Wayland,
not without patching every compositor in existence. In X11, enabling the support is
as easy as setting a xinput property (provided that a fork of X11 libinput driver is
used as well).

With three-finger drag, you lose three-finger swipe, understandably.

This also bumps version to 1.16.999 because it's ABI is not the same as in 1.15. 999 is there
so that upstream updates don't get to write their upgrade over, and not to spoil
changelog by clashes with upstream versioning.